Improve SuperPMI asserts
Include `GetIndex` method calls in the assert message, to distinguish
the different assert cases in error output. E.g., the asserts
change from:

```
'index != -1' failed ...
```

(which was identical for several cases), to:

```
'GetClassAttribs->GetIndex(CastHandle(classHandle)) != -1' failed ...
```

This makes it easier to know where the problem is without debugging the failure.