[wasm] Improve AOT builds (#48184)
* [wasm] pack: Let the workload configure the runtime pack path

* [wasm] update webassembly pack description

* [wasm] Makefile: fix being able to `make build`

* [wasm] Fix AOT compilation to work before publish

Instead of using runtime pack, and other assemblies from the
publish folder, we are now using them directly from the source
location. This means that `System.Private.CoreLib` can be in a different
directory, than the assembly being precompiled.

So, we need to set `MONO_PATH` to the relevant paths that will have the assemblies.

This manifests during blazor workload based builds.

* [wasm][aot] Don't strip the source assemblies in place

Currently we run `mono-cil-strip` on assemblies, in the AOT case, but we
do that in-place.

This is problematic because the assemblies from the runtime pack will
get modified in their original location (nuget/etc), since we don't use
them from the publish folder anymore.

- This would mean that a subsequent attempt to build with AOT would fail
  because now the original assemblies are already stripped!

- this also becomes apparent if you have the runtime pack installed in a
  system dotnet folder, eg. as a pack - `/usr/local/share/dotnet/packs/Microsoft.NETCore.App.Runtime.browser-wasm/`,
  because a regular user wouldn't have permissions to modify that.

Solution: Write the stripped assemblies into the obj dir

* [wasm] Make the final list of wasm assemblies as an item

- Blazor skips the generation of the app bundle when running
 `WasmBuildApp` target.
- Because of the previous commit, the locations for the final assemblies
  could be different (when aot+stripping), which needs to be surfaced
  for blazor. So, generate a `@(WasmAssembliesFinal)` item for that.

* [wasm] Allow enabling assert messages with $(MonoEnableAssertMessages)

* [wasm][aot] WasmApp.targets: check that we have .bc file for each assembly

Just a sanity check to ensure that due to some build authoring error, we
are not building dotnet.js/wasm with no .bc files! That just shows up as
an assert:

`* Assertion at /Users/radical/dev/r3/src/mono/mono/mini/aot-runtime.c:2651, condition `<disabled>' not met`

.. which is:
    `g_assert (info->version == MONO_AOT_FILE_VERSION);`

..  `0 != MONO_AOT_FILE_VERSION`

* [wasm] remove UsingTask for non-exitant WasmAOT

* MonoCrossAot pkg: use mono-aot-cross.exe on windows

* [wasm] Fix up emscripten command invocations for windows

* [wasm] WasmApp.targets: move some property inits closer to their use

* [wasm] fix build for browser-profile sample

* [wasm] fix emcc invocation for getting the version

* [wasm] Add new RunWithEmSdkEnv task

* [wasm] Use `RunWithEmSdkEnv` to run `emcc` commands

* [wasm] AotCompiler: use platform specific path separator

* [wasm] RunWithEmSdkEnv: quote path on windows

* [wasm] Use new ReadEmccVersion to get the versions

This simplifies the target too, and works on windows.

* Update src/installer/pkg/sfx/Microsoft.NETCore.App/Microsoft.NETCore.App.MonoCrossAOT.Sdk.props

Co-authored-by: Pranav K <prkrishn@hotmail.com>

* fix messageimportance for debug msg

* Fix emcc invocation on windows, thanks to @pranavkm

* Don't strip assemblies with mono-cil-strip for aotinterp mode

* [wasm] Remove ReadEmccVersion task, and use msbuild to get the version

.. string instead. We do this by looking for the line with `'emcc'` in
it, since the other lines are the license text. Hopefully, this isn't
too brittle, since we are invoking `emcc --version`.

* Fix emcc version extraction to not depend on a specific string

Co-authored-by: Pranav K <prkrishn@hotmail.com>