All bignums should be created as positive values.
All of the algorithms we support that use BigIntegers operate in modulus fields, where all negative values are congruent to a positive value. As a result, we don't use the negative values in the crypto implementations.

For example, explicitly specifying the NistP256 curve has a P, A, and Order that all start with FF in the byte array, but should be positive values.