Implement IgnoresAccessChecksToAttribute support
System.Runtime.CompilerServices.IgnoresAccessChecksToAttribute is like the reverse of
InternalsVisibleTo - it's added to an assembly to indicate that that
assembly can call non-public methods from the specified assembly.

The interesting thing is that the DynamicProxy dynamically generates
the attribute, so it is not explicitly declared in CoreLib.

Fixes https://github.com/dotnet/runtime/issues/47989