Improve the change based on Michal's PR feedback
Thanks to Michal's feedback I've been able to better grasp the
problem and fix deficiencies of my first commit. I have moved
setting of the constLookup to the "zapinfo version of getCallInfo",
I have modified the code to avoid canonicalization only for
array constructors and I rolled back the change to
CreateMethodEntrypoint that is no longer necessary after the
"zapinfo" fix.

Thanks

Tomas