Remove duplicated ILLink PackageReference and update target version of the SDK to 6.0 (#48462)
* Remove duplicated ILLink PackageReference

The Arcade.SDK already package refrences the ILLink package. The
duplicate reference in illink.targets caused SDK errors as the Arcade
reference has IsImplicitlyDefined set which doesn't allow an additional
reference with the same identity.

Also, as the ILLink package already exposes the path to the assembly via
its props file, using that instead of manually constructing the path to
the assembly.

The SDK target version update is required as the sequencing of the
ILLink.props file was wrong and is required for this change. This isn't
considered a breaking change, as the SDK's minimum required version
isn't changed.

* Update arcade dependencies

* Remove NuGet pack tasks pkgref

* Add mega hack workaround

* Remove KnownFrameworkReference items

* Don't hardcode SDK value in helix submission...

* Update runtimeConfiguration.targets

* Fix double publishing error in mobile tests

* Set DotNetCliVersion to right version for aspnetcoreruntime

* Update sendtohelixhelp.proj

* Update sendtohelixhelp.proj

Co-authored-by: Santiago Fernandez Madero <safern@microsoft.com>