Fix EventPipe source using COMPILE_OPTIONS on old CMake 3.6.2.
Older CMake < 3.11 doesn't include COMPILE_OPTIONS source file property.
This property is needed by clang or it will issue warning when building
files with .c extension under c++ compiler.

Fallback to add_compile_options if CMake is older than 3.11. Even if
that option has wider scope, this is only applied to files in eventpipe
directory or in debug-pal.

Keeping the more precise set_source_files_properties on newer CMake and
when we can upgrade CMake to 3.11 or newer we could get rid of this.