Fix majority of failures in libraries-jitstressregs after EHWriteThru enabled (#48829)
* Fix majority of libraries-jitstressregs

When we verify final allocation, we were not resetting the `assignedInterval` if the interval got assigned to a different register. This was not consistent to what we do during allocation.

Fixed following failures:

https://dev.azure.com/dnceng/public/_build/results?buildId=989595&view=ms.vss-test-web.build-test-results-tab&runId=30998546&resultId=169282&paneView=dotnet-dnceng.dnceng-build-release-tasks.helix-test-information-tab

https://helixre8s23ayyeko0k025g8.blob.core.windows.net/dotnet-runtime-refs-pull-47307-head-5f85fedec0cd4ff9a0/System.Net.Security.Tests/console.73e69194.log?sv=2019-07-07&se=2021-03-03T08%3A52%3A21Z&sr=c&sp=rl&sig=naQOrM%2BszKDyDAO8m%2BU%2Fu149f4cjodTk1dYCKl9cvKs%3D

Failures happen on arch/OS
- Linux arm64
- Linux arm
- Linux x64
- Windows x86
- Windows x64

* Delete regsToFree and delayRegsToFree from verifyFinalAllocation

These two variables were always set to `RBM_NONE`. As part of https://github.com/dotnet/runtime/issues/48837, investigate how to reintroduce them.

* jitformat