Add a test to verify current behavior of floating point (#48438)
The behavior of floating point to integer conversions outside of the valid range is not consistent today from platform to platform.

This test codifies the existing behavior, so that at the very least, any changes are understood.

As a summary, arm64 follows what is described as the saturating model, x86/x64 follow a model where 32/64 bit signed conversions have sentinel behavior (where NaN and out of range values are set to minint), and the unsigned conversion is performed via the signed 64bit conversion. And finally arm32 follows a model where 32bit operations are saturating, but 64bit conversions are defined in terms of the CRT 64 bit int signed conversion, which is defined based on the CRT 64 unsigned int conversion.

Implementations are provided in both C++ and C# using only standardized defined behavior. Both C++ and C# are used in order to detect cases where one or the other does not correctly implement conversions over the defined range.