Support loading ICU data from managed Interop
In iOS we support loading a custom dat file when working with ICU.  The way this worked originally was the mono runtime exported a function that
native code would call into (similar to wasm).  After thinking about it a bit, it makes more sense to load this the same way we do on desktop, but with the
ability to provide the path to an ICU dat file via an AppContext key `ICU_DAT_FILE_PATH`.  This can be provided before Xamarin iOS calls `monovm_initialize` and they won't have to worry about calling some special function.