Avoid boxing response content length in SocketsHttpHandler
The current header model means every Content-Length response results in us boxing a long and allocating a HeaderStoreItemInfo to track it.  And if headers are enumerated after that, we then end up regenerating a string from that boxed long.  The ContentLength property is generally only accessed once or twice, making it faster to just parse it every time.