Avoid forcing parsing of DefaultRequestHeaders
In the past, this was done because every request would parse all headers as part of sending them, so it was better to parse once then on every request.  But we changed the behavior for sending in SocketsHttpHandler to actually respect TryAddWithoutValidation, such that we wouldn't forcefully validate unless the developer asked us to.  This makes DefaultRequestHeaders consistent with that.