JIT: extend redundant branch opt to catch partially redundant cases
If the current relop has PHI inputs, see if any of those inputs would
produce the same relop VN as a dominating compare; if so the current
relop is partially redundant, and we may be able to optimize some of
the paths through the relop block via jump threading.

Addresses cases like the one seen in #48115, though that particular
case is not optimized as the current relop block has a side effect.