Various improvements
- Add support for the EmitDefaultMessages generator option which
controls whether default messages are produced if the user doesn't
supply one.

- Add support for the FieldName generator option which determines
the name of the field used to retrieve the logger instance when
using instance mode logging methods.

- Add support for the {OriginalFormat} element when enumerating the
TState. This is for compatibility with the existing LoggerMessage.Define
method.

- Add support for automatically expanding IEnumerable logging method
parameters to comma-separated strings. This is for compatibility
with the LoggerMessage.Define method.