Avoid forcing parsing of DefaultRequestHeaders and from message.ToString() (#49673)
* Avoid forcing parsing of DefaultRequestHeaders

In the past, this was done because every request would parse all headers as part of sending them, so it was better to parse once then on every request.  But we changed the behavior for sending in SocketsHttpHandler to actually respect TryAddWithoutValidation, such that we wouldn't forcefully validate unless the developer asked us to.  This makes DefaultRequestHeaders consistent with that.

* Avoid forcing parsing when ToString()'ing request and response messages

This not only can change behavior in the debugger just by hovering over a message, it can change behavior at runtime due to the request/response messages getting logged.

* Add another test