Propagate annotations over Type.BaseType (dotnet/linker#1901)
Some of the annotations apply with the same meaning on the base type - `All` and all of the `Public..`. It makes sense to propagate these over the call to `Type.BaseType`.

This change implements that. It also implements that if the input type is a known type the return value of `BaseType` is also know type - the actual base type of the type passed in.

Adds tests for both cases.

Co-authored-by: Michal Strehovský <MichalStrehovsky@users.noreply.github.com>

Commit migrated from https://github.com/dotnet/linker/commit/ed39b81a8ff5c4df3d93dcdf800f84e18622fed5