wip - components - big component loading hack - code runs now
the component loading hack is to look for the components in the directory of
the executable app.  This works for self-contained apps, but is wrong when
libmonosgen-2.0 is loaded by a host application.  We need to somehow find out
either from the host or from the system where that library _is_ and to load the
components from the same directory