Speed up SuperPMI
SuperPMI logging was doing a lot of work even if no log output was
asked for. For example, there are LogDebug() calls in the recordRelocation
path that were doing all the printf work and memory allocation even though
we were not logging anything.

Add a fast-path return to avoid all this unnecessary work.