Specify intermediate certificate for ChainTests.BuildInvalidSignatureTwice.
This test was relying on AIA certificate fetching to get the
intermediate certificate. This is not the purpose of this test,
and the fetching appears to have gotten unreliable.

This change uses a different certificate for the test and puts the
intermediate certificate in the extra store so that AIA fetching
is not used.