Resolve ILLink warnings in System.ComponentModel.TypeConverter (Round 3) (#49786)
* Resolve ILLink warnings in System.ComponentModel.TypeConverter (Round 3)

Contributes to #45623

* Mark GetProperties as RequiresUnreferencedCode.

PropertyDescriptors are not trim compatible because we can't guarantee the PropertyType is annotated as DynamicallyAccessedMembers.All.

* Mark GetConverter as RequiresUnreferencedCode.

NullableConverter requires the underlying type to be annotated with DynamicallyAccessedMembers.All, but it cannot be statically annotated.

* Mark GetDefaultEvent and GetDefaultProperty as RequiresUnreferencedCode.

GetDefaultProperty is not trim compatible because it calls GetProperties, which isn't trim compatible since the PropertyType can't be annotated.

GetDefaultEvent can technically be considered trim compatible when not passing in an object "instance" and only a Type. But this would require trim warning suppressions which aren't worth the risk since most of ICustomTypeDescriptor is already RequiresUnreferencedCode.