wip - put stubs in the component directory
For the static components scenario, build .a files for components and component
stubs and leave it to the final link step to pick what to link.

Also in the static components scenario, the stub implements a normal (ie
mono_component_<component_name>_init) entrypoint.

The advantage is that the embedder/workload doesn't need to write any
code (just pick libraries to link) and the runtime can do direct calls to the
component init functions avoiding having to use linker flags to ensure that
component entrypoints are in the final linked artifact.