JIT: fixes for mixed PGO/nonPGO compiles
Always scale inlinee counts, independent of the pgo status of the
call site and inlinee.

Get rid of `setBBWeight` and `modifyBBWeight` as they are attractive
nuisances that lead to bad habits. Use `inherit` and `scale` instead as
new counts should always be based on old.

Account for cases where there are methods in the inline tree with PGO
data but the root method doesn't have PGO, and vice versa.