[AppleAppBuilder] Entitlements to run tests on catalyst using the JIT (#50637)
* [catalyst] Add jit entitlements to AppleAppBuilder

to use the JIT on MacCatalyst we need the hardened runtime and the JIT
entitlement.  Otherwise mmap() with a MAP_JIT argument fails with EINVAL.

* [catalyst] also add disable library validation entitlement

To load libSystem.Native.dylib from the Resources/ directory in the .app
bundle.  (And possibly to load libicu from homebrew)

* [AppleAppBuilder] cleanup entitlements generation a little

Use a list in the builder instead of hardcoding in the template.

* [mono] update iOS sample to run on Catalyst too

use `make run-catalyst`

* fix typos and address feedback