Add MonoAOTCompiler MSBuild Task Workload
We were originally publishing the MonoAOTCompiler nuget and having Blazor, iOS, and Android include it as part of their workload.  Unfortunately, the workload spec does not allow multiple pack references from different manifests.  To get around this 1-1 relationship, we can supply a containing workload that can be consumed via the extends key.

Resolves https://github.com/dotnet/sdk/issues/16700