Fix resolution code for scenarios where values are copied from stack
During resolution, when we decide to resolve a `targetReg` "RA" which is the source of
another register "RB" that is also participating in resolution, we would spill "RA" and
add resolution move into "RA". However, we were not recording the fact that resolution of "RB"
now has to happen from `REG_STK`.

Also, when we perform resolution, we add back the source register into `targetRegsReady`
(data structure to tell which `targetReg`s are ready for resolution). This is not needed if
that source register was never one of the `targetReg`s we decided to resolve.