[System.Runtime.Loader] Add hot reload test infrastructure
Adding infrastructure for hot reload testing.

For each test we define a new library assembly project.  The `.csproj` has a
`DeltaScript` property that specifies a JSON file that lists the name of an
initial source file, and a list of updated versions of that file.  The
`hotreload-delta-gen` tool runs during the build to read the delta script and
create deltas that incorporate the updates.

The main testsuite references all the test assemblies, and when a test runs, it
calls `ApplyUpdateUtil.ApplyUpdate` to load subsequent deltas
and then compares the results before and after an update.

Dependencies:

- https://github.com/dotnet/hotreload-utils  the `hotreload-delta-gen` binary
must be installed and on the PATH  (TODO: package it as a dotnet tool and
publish to a transport nuget package)

Needs work:

- Mono test runs need to pass the `FEATURE_MONO_METADATA_UPDATE` property to
  msbuild to the test project.  This is because not every mono configuration
  includes hot reload support.  Eventually this should be subsumed by the
  runtime API to querty hot reload capabilities https://github.com/dotnet/runtime/issues/50111
- All runs need to pass `DOTNET_MODIFIABLE_ASSEMBLIES=debug` to the testsuite.
  Hot reload only works when:
  - the assemblies are compiled with the Debug compiler setting,
  - and, if the runtime is tarted with the above environment variable set.
- The GenerateHotReloadDeltas.targets needs some output file work to run `hotreload-delta-gen` less
  frequently.  Right now it runs every time even if everything is up to date.
  For CI testing we should ideally compile the deltas ahead of time once.

To try it out locally:

1. checkout and build `hotreload-utils` and do `dotnet publish --self-contained -r <RID>` to
put `hotreload-delta-gen` into that projects' `artifacts/...` forlder.  Add the
publish folder to your `$PATH`

2. In dotnet/runtime run
```
DOTNET_MODIFIABLE_ASSEMBLIES=debug MONO_ENV_OPTIONS=--interp ./dotnet.sh build src/libraries/System.Runtime.Loader/tests /p:MonoMetadataUpdate=true /t:Test
```

(CoreCLR doesn't need `MONO_ENV_OPTIONS` or `MonoMetadataUpdate`)